pyproject.toml
[tool.poetry]
name = "audio-search-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "audio_search_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
yt-dlp = "^2024.3.10"
moviepy = "^1.0.3"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
python-dotenv = "^1.0.1"
sentence-transformers = "^2.7.0"
pandas = "^2.2.2"
chromadb = "^0.5.0"
qdrant-client = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
